% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCtagging.R
\name{absolutely.in.area}
\alias{absolutely.in.area}
\title{absolutely.in.area}
\usage{
absolutely.in.area(area, abslon, abslat)
}
\arguments{
\item{area}{The area name exa. 'cfa23' 'cfa24' 'nens' 'sens' 'gulf' 'cfa4x'}

\item{lon}{The longitude of position in question}

\item{lat}{The latitude of position in question}
}
\value{
TRUE or FALSE
}
\description{
Function that determins if a position is inside defined polygons
}
