% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCtagging.R
\name{absolutely.in.area2}
\alias{absolutely.in.area2}
\title{absolutely.in.area2}
\usage{
absolutely.in.area2(area, abslon, abslat)
}
\arguments{
\item{area}{The area name exa. 'cfa23' 'cfa24' 'nens' 'sens' 'gulf' 'cfa4x'}

\item{lon}{The longitude list}

\item{lat}{The latitude list}
}
\value{
list of TRUE or FALSE values
}
\description{
Function that determines if a list of positions are inside defined polygons
}
