% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagWebGUI.r
\name{autoaddData}
\alias{autoaddData}
\title{autoaddData}
\usage{
autoaddData(name = "", region = "")
}
\arguments{
\item{name}{The persons name from which to obtain current info}
}
\description{
Function that help autopopulate people in the html form
}
