% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAccoustic.R
\name{compress.detections}
\alias{compress.detections}
\title{compress.detections}
\usage{
compress.detections(hours.thres = 3)
}
\value{
dataframe
}
\description{
Simplifies the data by grouping detections.
}
