% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCtagging.R
\name{create.tag.geojson}
\alias{create.tag.geojson}
\title{create.tag.geojson}
\usage{
create.tag.geojson(filename = NULL)
}
\arguments{
\item{filename}{path to write the resulting kml}
}
\description{
Write tag movement data to GeoJSON javascript functions. This is used by the enssnowcrab website to display tag movement data by year. The previous tag kml does not display properly in google maps so this method was implemented. It looks better than the old kml since geoJSON has built in arrow markers.
}
