% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCtagging.R
\name{create.tag.kml.2}
\alias{create.tag.kml.2}
\title{create.tag.kml.2}
\usage{
create.tag.kml.2(write = T, preview = T, tofile = NULL)
}
\arguments{
\item{filename}{path to write the resulting kml}
}
\description{
Write tag movement data to kml.
}
