% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCtagging.R
\name{create.tag.kml}
\alias{create.tag.kml}
\title{create.tag.kml}
\usage{
create.tag.kml(preview = T, write = T, tofile = NULL)
}
\arguments{
\item{preview}{Boolean default TRUE}

\item{write}{Boolean default TRUE}

\item{tofile}{path to write the resulting kml}
}
\description{
Function that generates a kml plot of tag data
}
