% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCtagging.R
\name{degmin2decdeg}
\alias{degmin2decdeg}
\title{degminsec2decdeg}
\usage{
degmin2decdeg(ddmmss.ss)
}
\value{
numeric converted coordinate
}
\description{
Function that converts degree minutes decimal seconds to decimal degrees
}
