% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAccoustic.R
\name{get.acoustic.detections}
\alias{get.acoustic.detections}
\title{get.acoustic.detections}
\usage{
get.acoustic.detections(region = "ScotianShelf")
}
\value{
dataframe
}
\description{
Get accoustic detection data from Oracle Database
}
