% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAccoustic.R
\name{get.acoustic.releases}
\alias{get.acoustic.releases}
\title{get.acoustic.releases}
\usage{
get.acoustic.releases(region = "ScotianShelf")
}
\value{
dataframe
}
\description{
Write release data to Oracle Database
}
