% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCtagging.R
\name{get.capturedata.oracle}
\alias{get.capturedata.oracle}
\title{get.capturedata.oracle}
\usage{
get.capturedata.oracle()
}
\value{
dataframe of data
}
\description{
Get all tag data view from oracle database
}
