% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCtagging.R
\name{open.port.SC}
\alias{open.port.SC}
\title{open.port.SC}
\usage{
\method{open}{port.SC}(user = enssnowc.user, password = enssnowc.password,
  host = "enssnowcrab.com", local.port = NULL, remote.port = "3306")
}
\arguments{
\item{user}{Database username}

\item{password}{Database password}

\item{host}{database server}

\item{local.port}{what port to tunnel through}

\item{remote.port}{what port to tunnel through}
}
\description{
Open a ssh tunnel for ENS site satabase opperations
}
