% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagWebGUI.r
\name{SC_Stats_Capture}
\alias{SC_Stats_Capture}
\title{SC_Stats_Capture}
\usage{
SC_Stats_Capture(are = "", years = "", region = "SoctianShelf")
}
\value{
message to webpage
}
\description{
Function generates stats for display on webpage
}
