% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagWebGUI.r
\name{SC_Stats_Capture_Table}
\alias{SC_Stats_Capture_Table}
\title{SC_Stats_Capture_Table}
\usage{
SC_Stats_Capture_Table(are = "", years = "", region = "SoctianShelf")
}
\value{
message to webpage
}
\description{
Function generates stats in Table format for display on webpage
}
