% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagWebGUI.r
\name{SC_Stats_Sample}
\alias{SC_Stats_Sample}
\title{SC_Stats_Sample}
\usage{
SC_Stats_Sample(area = "", years = "", region = "SoctianShelf")
}
\value{
message to webpage
}
\description{
Function generates stats for display on webpage
}
