% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagWebGUI.r
\name{SCT_nrows}
\alias{SCT_nrows}
\title{SCT_nrows}
\usage{
SCT_nrows(table = "")
}
\arguments{
\item{table}{The name of the table from which to determine the current number of rows}
}
\description{
Get the number of rows from specified table
}
