\name{SCtagging-package}
\alias{SCtagging-package}
\alias{SCtagging}
\docType{package}
\title{
\packageTitle{SCtagging}
}
\description{
\packageDescription{SCtagging}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SCtagging}
\packageIndices{SCtagging}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SCtagging}

Maintainer: \packageMaintainer{SCtagging}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
