% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCtagging.R
\name{shortestpaths.SC}
\alias{shortestpaths.SC}
\title{Shortest Paths}
\usage{
shortestpaths.SC(raster.path = system.file("extdata", "depthraster2.tif",
  package = "SCtagging"), neighborhood = 16, type = "random.walk",
  redo = F, region = "ScotianShelf")
}
\arguments{
\item{neighborhood}{The number of adjacent cell over which to calculate, defaults to 16.}

\item{type}{The type of calculation, either 'random.walk' or 'least.cost'. Defaults to random.walk}

\item{redo}{Set redo = TRUE if you want to rewrite all data, FALSE to only update new entries}

\item{region}{Either 'ScotianShelf' or 'Gulf'}

\item{rasterpath}{The canvas on which to calculate shortest path, defaults to depth raster.}
}
\value{
dataframe
}
\description{
Creates and writes the shortest paths to database
}
