% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagWebGUI.r
\name{tagReturned_Applied}
\alias{tagReturned_Applied}
\title{tagReturned_Applied}
\usage{
tagReturned_Applied(are, years, rm.gulf = T)
}
\value{
stats message
}
\description{
Helper function to generate stats
}
