% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagWebGUI.r
\name{tagReturned_Year}
\alias{tagReturned_Year}
\title{tagReturned_Year}
\usage{
tagReturned_Year(are, years, region = "SoctianShelf")
}
\value{
stats message
}
\description{
Helper function to generate stats
}
