% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAccoustic.R
\name{write.accoustic.paths}
\alias{write.accoustic.paths}
\title{write.acoustic.detections}
\usage{
write.accoustic.paths(x, rel)
}
\value{
dataframe
}
\description{
Write calculated path data to Oracle Database
}
