% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAccoustic.R
\name{write.acoustic.detections}
\alias{write.acoustic.detections}
\title{write.acoustic.detections}
\usage{
write.acoustic.detections(dir = file.path("E:", "OTN", "detections"))
}
\value{
dataframe
}
\description{
Write release data to Oracle Database
}
