% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAccoustic.R
\name{write.acoustic.releases}
\alias{write.acoustic.releases}
\title{write.acoustic.releases}
\usage{
write.acoustic.releases(dir = file.path("E:", "OTN", "releases"))
}
\value{
dataframe
}
\description{
Write release data to Oracle Database
}
