% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCtagging.R
\name{mirror2esnssite}
\alias{mirror2esnssite}
\title{mirror2esnssite}
\usage{
mirror2esnssite(region = "ScotianShelf")
}
\value{
status message in case it was called by webpage
}
\description{
ESSENTIAL FUNCTION. Must call after any data entry in order for the website to be up to date!
}
